save before cleanup
