+++ scripts/Validation/run_validation.sh
@@ -1,7 +1,7 @@
 infile="$1"
-n="$2"
+shift
 base_filename=$(basename "$infile")
 outfile="/exp/dune/data/users/${USER}/dune-tms/Validation/Tracking_Validation/${base_filename}"
-make && ./Tracking_Validation "$infile" $n && python simply_draw_everything.py "${outfile}"
+make && ./Tracking_Validation "$infile" $@ && python simply_draw_everything.py "${outfile}"
 echo "Output should be in"
 echo "${outfile/.root/_images}"
